#971 [REFACTOR] GuideModal shared 컴포넌트로 수정
