+++ src/feature/event/component/index.js
@@ -1,3 +1,3 @@
 export { default as EventForm } from './EventForm/EventForm';
 export { default as NoticeForm } from './NoticeForm/NoticeForm';
-export { default as GuideModal } from './GuideModal/GuideModal';
+export { default as GuideModal } from '../../../shared/component/modal/GuideModal/GuideModal';
