+++ src/shared/constant/modalOptions.js
@@ -3,6 +3,11 @@ import {
   POST_REPORT_TYPE_MODAL_OPTION_CHILDREN,
   USER_REPORT_TYPE_MODAL_OPTION_CHILDREN,
 } from './board';
+import {
+  noParticipationIllustration,
+  noSellIllustration,
+  noTransferIllustration,
+} from '@/assets/illustrations';
 
 export const MODAL_OPTIONS = [
   {
@@ -219,3 +224,39 @@ export const MODAL_CONFIRM = {
     ),
   },
 };
+
+export const EVENT_GUIDE_MODAL_OPTIONS = [
+  {
+    title: '무단 불참 절대 금지!',
+    content: (
+      <>
+        무단 불참 시 경고 1회가 부여되며,
+        <br />
+        향후 모든 이벤트에 참여가 불가해요
+      </>
+    ),
+    image: noParticipationIllustration,
+  },
+  {
+    title: '양도, 대리수령 절대 금지!',
+    content: (
+      <>
+        양도,대리수령 적발 시 2년 강등되며,
+        <br />
+        향후 모든 이벤트에 참여가 불가해요
+      </>
+    ),
+    image: noTransferIllustration,
+  },
+  {
+    title: '입장권 금전 거래 절대 금지!',
+    content: (
+      <>
+        입장권 금전 거래 적발 시,
+        <br />
+        영구 강등 처리돼요
+      </>
+    ),
+    image: noSellIllustration,
+  },
+];
